Remove constexpr from composite_mask.
Having constexpr leads to the whole array to be computed
at compile time. This might be nice for some production code,
where having this pre-built would speed things up, but for most
applications this is just bad. And it takes for ever to compile.
Much faster to build at runtime.